backend/requirements.txt
django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
dj-database-url==2.2.0
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.1.1
django-celery-beat==2.7.0
gunicorn==23.0.0
anthropic==0.39.0
playwright==1.48.0
beautifulsoup4==4.12.3
Pillow==10.4.0
python-decouple==3.8
